v6.6.9 - Big Restructuring
The start of a big revamp to global REMs over px's - I'm pushing it to GIT at this point as now I'm diving into each section of the site one at a time. So in case I mess something up, I can roll back to this working point with new folder structure.